atomic-5.C (f1): Adjust expected lines of read-only variable messages.
	* g++.dg/gomp/atomic-5.C (f1): Adjust expected lines of read-only
	variable messages.

	* testsuite/libgomp.c-c++-common/taskloop-reduction-1.c (S): Change
	type of s and t members from unsigned long int to
	unsigned long long int.

From-SVN: r265887

+++ libgomp/testsuite/libgomp.c-c++-common/taskloop-reduction-1.c
@@ -4,7 +4,7 @@ extern
 #endif
 void abort (void);
 
-struct S { unsigned long int s, t; };
+struct S { unsigned long long int s, t; };
 
 void
 rbar (struct S *p, struct S *o)
